+++ src/components/Main/Metrics/MetricsCompany/MetricsCompany.jsx
@@ -5,12 +5,29 @@ import $MetricsCompany from "./MetricsCompany.styles.jsx";
 const tempCompany = [
   {
     id: 1,
-    name: "ING Hubs Polska",
+    name: "ING Hubs",
     postal_code: "12-123",
     street: "Zabrska",
     street_number: 17,
     apartment_number: 1,
-    carbon_limit: 700.0,
+    carbon_limit: 100.0,
+    location: {
+      id: 1,
+      city: "Katowice",
+      country: {
+        id: 3,
+        name: "Polska",
+      },
+    },
+  },
+  {
+    id: 5,
+    name: "ING Hubs",
+    postal_code: "12-123",
+    street: "Zabrska",
+    street_number: 17,
+    apartment_number: 1,
+    carbon_limit: 100.0,
     location: {
       id: 1,
       city: "Katowice",
@@ -22,35 +39,69 @@ const tempCompany = [
   },
   {
     id: 2,
-    name: "ING Hubs Romania",
+    name: "ING Hubs",
     postal_code: "12-124",
     street: "Kultury",
     street_number: 232,
     apartment_number: 32,
-    carbon_limit: 750.0,
+    carbon_limit: 95.0,
     location: {
       id: 3,
       city: "Warszawa",
       country: {
         id: 4,
-        name: "Polska",
+        name: "Romania",
       },
     },
   },
   {
-    id: 4,
-    name: "ING Hubs Philipines",
+    id: 3,
+    name: "ING Hubs",
     postal_code: "12-124",
     street: "Zeusa",
     street_number: 232,
     apartment_number: 32,
-    carbon_limit: 100.0,
+    carbon_limit: 50.0,
     location: {
       id: 3,
       city: "Roma",
       country: {
         id: 4,
         name: "Italy",
+      },
+    },
+  },
+  {
+    id: 6,
+    name: "ING Hubs",
+    postal_code: "12-123",
+    street: "Zabrska",
+    street_number: 17,
+    apartment_number: 1,
+    carbon_limit: 100.0,
+    location: {
+      id: 1,
+      city: "Katowice",
+      country: {
+        id: 3,
+        name: "Polska",
+      },
+    },
+  },
+  {
+    id: 7,
+    name: "ING Hubs",
+    postal_code: "12-123",
+    street: "Zabrska",
+    street_number: 17,
+    apartment_number: 1,
+    carbon_limit: 1.0,
+    location: {
+      id: 1,
+      city: "Katowice",
+      country: {
+        id: 3,
+        name: "Polska",
       },
     },
   },
@@ -65,17 +116,17 @@ export default function MetricsCompany() {
     setAvailableCompanies(companies);
   }, []);
 
-  const currentUsage = (id) => {
+  function currentUsage(id) {
     // TODO api call to get numebr of areas where company.id = id
     // na razei mock, powinno byc jako useEffect
     // chyba jakis useEffect
-    return 550;
+    return 90;
   }
 
   return (
     <$MetricsCompany>
       {availableCompanies.map((company, index) => {
-        return <CompanyItem key={index} company={company} currentUsage={550} index={index}/>;
+        return <CompanyItem key={index} company={company} currentUsage={currentUsage(company.id)} index={index}/>;
       })}
     </$MetricsCompany>
   );